Use nubs/which to locate commands in user's PATH. Fixes #4.
Hardcoded paths are not very friendly as they may break on many systems
(e.g., [NixOS]).  Instead, allow the user to pass a command locator that
can be used to find a command anywhere in the configured PATH.

nubs/which has several ways of specifying the PATH, so it can be used in
practically any *nix based situation.

[NixOS]: http://nixos.org/