Enable the default editor to be a list of editors.
In general, /bin/ed does not make a good fallback editor.  Very few
people know how to use it.  specifying some other editors to also check
for helps make this more functional even for those who don't have a
proper EDITOR environment variable set.

I chose nano and vim to be tested before ed.  nano is an editor that
almost anyone can at least use so it makes sense as a first check.
Those who don't like nano generally know enough to be able to set their
EDITOR environment variable to something else they would prefer.

The last editor in the list is chosen as the final fallback.  If no
editors can be found using the locator, then the least preferable, but
most likely to exist, editor makes the most sense to return.  This
definitely works for the case where the user does not include the
command locator.  In this case, we can't do much to verify the commands,
so just returning the last one is about as good as we can do.

I'm beginning to lean towards making the locator an actual hard
dependency.  The lack of functionality when you don't provide one is
definitely limiting.