Require usage of nubs/which locators.
These locators make things so much simpler to reason with and improve
usage all around.  The nicety around not having an external dependency
doesn't make a big enough difference when the functionality it provides
is so useful.

Perhaps in the future this dependency can be removed and made optional
again.